It's not gosass, it's now sassy
First off, don't build libsass as a shared library, move it into the
repository. This should make it easier to build the library using go
get. Then I had to remove the sass2scss library since it was licensed
under GPLv3 and I didn't want to dictate the license of users of this
library. Then I rewrote the go API since it was kind of... not Go-like
at all.

Reasoning behind this is that I wanted a sass library that I could use
easily and point people to, but the current sass libraries all required
you to build libsass, which was silly. Then when I was working through
the licensing situation, I realized that there was a random GPL library
in libsass, I'll write a Go version that isn't GPL'd but at the moment I
just don't need this junk.